+++ docs/CHANGELOG.md
@@ -1,4 +1,4 @@
-Version 3.34.0 (May 1, 2023)
+Version 3.34.0 (May 2, 2023)
 ------------------------------
 
 **User-visible changes:**